add ByteVector

+++ Tests/ImageCompression/Utils/Objects/DataOPCTest.kt
@@ -26,9 +26,9 @@ class DataOPCTest {
         val base=dopc.FromBaseToArray()
         val sing=dopc.FromSignToArray()
 
-        dopc.FromArrayToBigInt(kotlin.ShortArray(SIZEOFBLOCK,{x->0}))
-        dopc.FromArrayToBase(kotlin.ShortArray(SIZEOFBLOCK,{x->0}))
-        dopc.FromArrayToSing(kotlin.ShortArray(SIZEOFBLOCK,{x->0}))
+        dopc.FromArrayToBigInt(kotlin.ByteArray(SIZEOFBLOCK,{x->0}))
+        dopc.FromArrayToBase(kotlin.ByteArray(SIZEOFBLOCK,{x->0}))
+        dopc.FromArrayToSing(kotlin.ByteArray(SIZEOFBLOCK,{x->0}))
         assertFails { AssertDataOpcEqual(dopc,dopc1) }
 
         dopc.FromArrayToBigInt(bs)
